+++ spec/graphlient/static_client_query_spec.rb
@@ -15,7 +15,16 @@ describe Graphlient::Client do
         end
       end
 
-      Query = Client.parse do
+      StringQuery = Client.parse <<~GRAPHQL
+        query($some_id: Int) {
+          invoice(id: $some_id) {
+            id
+            feeInCents
+          }
+        }
+      GRAPHQL
+
+      BlockQuery = Client.parse do
         query(some_id: :int) do
           invoice(id: :some_id) do
             id
@@ -29,11 +38,31 @@ describe Graphlient::Client do
       expect(Graphlient::Client::Spec::Client.send(:client).allow_dynamic_queries).to be false
     end
 
-    it '#execute' do
-      response = Graphlient::Client::Spec::Client.execute(Graphlient::Client::Spec::Query, some_id: 42)
-      invoice = response.data.invoice
-      expect(invoice.id).to eq '42'
-      expect(invoice.fee_in_cents).to eq 20_000
+    context 'with string-based queries' do
+      it 'parses a string query to an OperationDefinition' do
+        expect(Graphlient::Client::Spec::StringQuery.class).to be GraphQL::Client::OperationDefinition
+      end
+
+      it 'sets the OperationDefinition that came from a string to have a name' do
+        expect(Graphlient::Client::Spec::StringQuery.definition_name).to eql 'Graphlient__Client__Spec__StringQuery'
+      end
+    end
+
+    context 'with both string- and block-based queries' do
+      it 'gets identical results parsing equivalent string- and block-based queries' do
+        block_response = Graphlient::Client::Spec::Client.execute(Graphlient::Client::Spec::BlockQuery, some_id: 42)
+        string_response = Graphlient::Client::Spec::Client.execute(Graphlient::Client::Spec::StringQuery, some_id: 42)
+        expect(string_response.to_h).to eq block_response.to_h
+      end
+    end
+
+    context 'executing a query' do
+      it 'succeeds with expected feeInCents' do
+        response = Graphlient::Client::Spec::Client.execute(Graphlient::Client::Spec::BlockQuery, some_id: 42)
+        invoice = response.data.invoice
+        expect(invoice.id).to eq '42'
+        expect(invoice.fee_in_cents).to eq 20_000
+      end
     end
   end
 end